feat(logging): modularize embed creation
