+++ src/commands/moderation/anti-raid-nuke.ts
@@ -9,8 +9,8 @@ import { inject, injectable } from 'tsyringe';
 import type { Command } from '../../Command.js';
 import { DATE_FORMAT_LOGFILE } from '../../Constants.js';
 import { type Case, CaseAction, createCase } from '../../functions/cases/createCase.js';
-import { generateCasePayload } from '../../functions/logs/generateCasePayload.js';
-import { insertAntiRaidNukeCaseLog } from '../../functions/logs/insertAntiRaidNukeCaseLog.js';
+import { generateCasePayload } from '../../functions/logging/generateCasePayload.js';
+import { insertAntiRaidNukeCaseLog } from '../../functions/logging/insertAntiRaidNukeCaseLog.js';
 import { checkModRole } from '../../functions/permissions/checkModRole.js';
 import { checkLogChannel } from '../../functions/settings/checkLogChannel.js';
 import { getGuildSetting, SettingsKeys } from '../../functions/settings/getGuildSetting.js';
