+++ src/commands/moderation/ban.ts
@@ -5,8 +5,8 @@ import { nanoid } from 'nanoid';
 import { inject, injectable } from 'tsyringe';
 import type { Command } from '../../Command.js';
 import { CaseAction, createCase } from '../../functions/cases/createCase.js';
-import { generateCasePayload } from '../../functions/logs/generateCasePayload.js';
-import { upsertCaseLog } from '../../functions/logs/upsertCaseLog.js';
+import { generateCasePayload } from '../../functions/logging/generateCasePayload.js';
+import { upsertCaseLog } from '../../functions/logging/upsertCaseLog.js';
 import { checkModRole } from '../../functions/permissions/checkModRole.js';
 import { checkLogChannel } from '../../functions/settings/checkLogChannel.js';
 import { getGuildSetting, SettingsKeys } from '../../functions/settings/getGuildSetting.js';
