+++ src/commands/moderation/duration.ts
@@ -4,7 +4,7 @@ import i18next from 'i18next';
 import type { Command } from '../../Command.js';
 import { getCase } from '../../functions/cases/getCase.js';
 import { updateCase } from '../../functions/cases/updateCase.js';
-import { upsertCaseLog } from '../../functions/logs/upsertCaseLog.js';
+import { upsertCaseLog } from '../../functions/logging/upsertCaseLog.js';
 import { checkModRole } from '../../functions/permissions/checkModRole.js';
 import { checkLogChannel } from '../../functions/settings/checkLogChannel.js';
 import { getGuildSetting, SettingsKeys } from '../../functions/settings/getGuildSetting.js';
