+++ src/commands/moderation/reason.ts
@@ -5,7 +5,7 @@ import type { Command } from '../../Command.js';
 import type { Case } from '../../functions/cases/createCase.js';
 import { getCase } from '../../functions/cases/getCase.js';
 import { updateCase } from '../../functions/cases/updateCase.js';
-import { upsertCaseLog } from '../../functions/logs/upsertCaseLog.js';
+import { upsertCaseLog } from '../../functions/logging/upsertCaseLog.js';
 import { checkModRole } from '../../functions/permissions/checkModRole.js';
 import { checkLogChannel } from '../../functions/settings/checkLogChannel.js';
 import { getGuildSetting, SettingsKeys } from '../../functions/settings/getGuildSetting.js';
