+++ src/commands/moderation/sub/restrict/react.ts
@@ -11,8 +11,8 @@ import { nanoid } from 'nanoid';
 import type { Sql } from 'postgres';
 import { container } from 'tsyringe';
 import { CaseAction, createCase } from '../../../../functions/cases/createCase.js';
-import { generateCasePayload } from '../../../../functions/logs/generateCasePayload.js';
-import { upsertCaseLog } from '../../../../functions/logs/upsertCaseLog.js';
+import { generateCasePayload } from '../../../../functions/logging/generateCasePayload.js';
+import { upsertCaseLog } from '../../../../functions/logging/upsertCaseLog.js';
 import type { ArgumentsOf } from '../../../../interactions/ArgumentsOf.js';
 import type { RestrictCommand } from '../../../../interactions/index.js';
 import { kSQL } from '../../../../tokens.js';
