+++ src/events/manual-ban/guildBanRemove.ts
@@ -6,7 +6,7 @@ import type { Redis } from 'ioredis';
 import { inject, injectable } from 'tsyringe';
 import type { Event } from '../../Event.js';
 import { deleteCase } from '../../functions/cases/deleteCase.js';
-import { upsertCaseLog } from '../../functions/logs/upsertCaseLog.js';
+import { upsertCaseLog } from '../../functions/logging/upsertCaseLog.js';
 import { checkLogChannel } from '../../functions/settings/checkLogChannel.js';
 import { getGuildSetting, SettingsKeys } from '../../functions/settings/getGuildSetting.js';
 import { logger } from '../../logger.js';
