+++ src/events/manual-kick/guildMemberRemove.ts
@@ -6,8 +6,8 @@ import type { Redis } from 'ioredis';
 import { inject, injectable } from 'tsyringe';
 import type { Event } from '../../Event.js';
 import { createCase, CaseAction } from '../../functions/cases/createCase.js';
-import { generateCasePayload } from '../../functions/logs/generateCasePayload.js';
-import { upsertCaseLog } from '../../functions/logs/upsertCaseLog.js';
+import { generateCasePayload } from '../../functions/logging/generateCasePayload.js';
+import { upsertCaseLog } from '../../functions/logging/upsertCaseLog.js';
 import { checkLogChannel } from '../../functions/settings/checkLogChannel.js';
 import { getGuildSetting, SettingsKeys } from '../../functions/settings/getGuildSetting.js';
 import { logger } from '../../logger.js';
