+++ src/functions/logging/generateAntiRaidNukeCaseLog.ts
