+++ src/functions/logging/generateCasePayload.ts
